create basic hashlynx app
