+++ hashlynx/public/manifest.json
