+++ hashlynx/src/App.js
@@ -1,0 +1,42 @@
+import * as Callbridge from '@iotum/callbridge-js';
+import React, { useState, useRef, useEffect } from 'react';
+import styles from './submitForm.module.css';
+
+function App() {
+  const [token, setToken] = useState('');
+  const [hostId, setHostId] = useState('');
+  const [submitted, setSubmitted] = useState(false);
+  
+  const container = useRef(null);
+
+  const handleSubmit = () => {
+    setSubmitted(true);
+  };
+
+
+  if (submitted) {
+    return (
+      
+    );
+  }
+
+  return (
+    <div className="form-wrapper">
+      <form onSubmit={handleSubmit}>
+        <label>
+          SSO Token:
+          <input type="text" value={token} onChange={handleTokenChange} />
+        </label>
+        <br />
+        <label>
+          Host ID:
+          <input type="text" value={hostId} onChange={handleHostIdChange} />
+        </label>
+        <br />
+        <button type="submit">Submit</button>
+      </form>
+    </div>
+  );
+}
+
+export default App;
